move version helper into server/ so it ships in the package
